remove extra details file
